ED: fixed code review comments.
