+++ DeanerySystem/DeanerySystem.Domain/Configurations/ProfessorConfiguration.cs
@@ -18,10 +18,8 @@ namespace DeanerySystem.Domain.Configurations
 			this.Property(p => p.Id).HasDatabaseGeneratedOption(DatabaseGeneratedOption.Identity);
 
 			this.HasRequired(p => p.Department).WithMany(d => d.Professors);
-
 			this.HasOptional(p => p.RectorOfUniversity).WithOptionalPrincipal(u => u.Rector);
 			this.HasOptional(p => p.DeanOfFaculty).WithOptionalPrincipal(f => f.Dean);
-			//this.HasOptional(p => p.HeadOfDepartment).WithOptionalPrincipal(d => d.Head);
 			this.HasOptional(p => p.MentorOfGroup).WithRequired(g => g.Mentor);
 
 			this.HasMany(p => p.Classes).WithRequired(j => j.Professor);
