+++ DeanerySystem/DeanerySystem.UI/DeanerySystem.UI.csproj
@@ -217,6 +217,7 @@
     </Compile>
     <Compile Include="Infrastructure\FeederFactory.cs" />
     <Compile Include="Infrastructure\MockFactory.cs" />
+    <Compile Include="Infrastructure\MockTesting.cs" />
     <Compile Include="Infrastructure\NinjectDependencyResolver.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
     <Compile Include="Models\Admin\Faculty.cs" />
